+++ internal/completions/httpapi/chat.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/sourcegraph/log"
 
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/envvar"
 	"github.com/sourcegraph/sourcegraph/internal/completions/types"
 	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 	"github.com/sourcegraph/sourcegraph/internal/database"
@@ -24,6 +25,9 @@ func NewChatCompletionsStreamHandler(logger log.Logger, db database.DB) http.Han
 		rl,
 		"chat",
 		func(requestParams types.CodyCompletionRequestParameters, c *conftypes.CompletionsConfig) (string, error) {
+			if isAllowedCustomChatModel(requestParams.Model) {
+				return requestParams.Model, nil
+			}
 			// No user defined models for now.
 			if requestParams.Fast {
 				return c.FastChatModel, nil
@@ -31,4 +35,22 @@ func NewChatCompletionsStreamHandler(logger log.Logger, db database.DB) http.Han
 			return c.ChatModel, nil
 		},
 	)
+}
+
+// We only allow dotcom clients to select a custom chat model and maintain an allowlist for which
+// custom values we support
+func isAllowedCustomChatModel(model string) bool {
+	if !(envvar.SourcegraphDotComMode()) {
+		return false
+	}
+
+	switch model {
+	case "anthropic/claude-2",
+		"anthropic/claude-instant-1.2-cyan",
+		"openai/gpt-3.5-turbo",
+		"openai/gpt-4-1106-preview":
+		return true
+	}
+
+	return false
 }